# feat: improve IsMethod 