--- a/header_timing_test.go
+++ b/header_timing_test.go
@@ -179,72 +179,11 @@ func BenchmarkRemoveNewLines(b *testing.B) {
 	}
 }
 
-func BenchmarkRequestHeaderIsPost(b *testing.B) {
-	req := &RequestHeader{method: strPost}
+func BenchmarkRequestHeaderIsGet(b *testing.B) {
+	req := &RequestHeader{method: []byte(MethodGet)}
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			req.IsPost()
-		}
-	})
-}
-func BenchmarkRequestHeaderIsHead(b *testing.B) {
-	req := &RequestHeader{method: strHead}
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			req.IsHead()
-		}
-	})
-}
-func BenchmarkRequestHeaderIsPut(b *testing.B) {
-	req := &RequestHeader{method: strPut}
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			req.IsPut()
-		}
-	})
-}
-
-func BenchmarkRequestHeaderIsDelete(b *testing.B) {
-	req := &RequestHeader{method: strDelete}
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			req.IsDelete()
-		}
-	})
-}
-
-func BenchmarkRequestHeaderIsConnect(b *testing.B) {
-	req := &RequestHeader{method: strConnect}
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			req.IsConnect()
-		}
-	})
-}
-
-func BenchmarkRequestHeaderIsOptions(b *testing.B) {
-	req := &RequestHeader{method: strOptions}
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			req.IsOptions()
-		}
-	})
-}
-
-func BenchmarkRequestHeaderIsTrace(b *testing.B) {
-	req := &RequestHeader{method: strTrace}
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			req.IsTrace()
-		}
-	})
-}
-
-func BenchmarkRequestHeaderIsPatch(b *testing.B) {
-	req := &RequestHeader{method: strPatch}
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			req.IsPatch()
+			req.IsGet()
 		}
 	})
 }