improve `IsMethod`  i.e `IsPost` 、 `IsGet ` and  so on.

benchmark
- old
```
goos: linux
goarch: amd64
pkg: github.com/valyala/fasthttp
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
BenchmarkRequestHeaderIsPost-2          559623241                2.003 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          604524886                1.992 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          601059054                2.015 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          605856414                2.006 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          606804846                1.981 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          608144640                1.984 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          606334010                2.000 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          606591694                2.001 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          606142994                2.007 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          602808153                1.974 ns/op           0 B/op          0 allocs/op
PASS
ok      github.com/valyala/fasthttp     14.092s
```

- new
```
goos: linux
goarch: amd64
pkg: github.com/valyala/fasthttp
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
BenchmarkRequestHeaderIsPost-2          982056084                1.123 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.127 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.121 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.122 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.114 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.132 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.114 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.115 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.111 ns/op           0 B/op          0 allocs/op
BenchmarkRequestHeaderIsPost-2          1000000000               1.133 ns/op           0 B/op          0 allocs/op
PASS
ok      github.com/valyala/fasthttp     12.406s

```